app/requirements.txt
Flask==3.0.0
pyodbc==5.0.1
azure-identity==1.15.0